+++ app/utils.py
@@ -1,6 +1,18 @@
+from decimal import Decimal
+
 from slugify import Slugify
+from connexion.decorators.produces import JSONEncoder
+
 
 custom_slugify = Slugify(to_lower=True)
+
+
+class DecimalEncoder(JSONEncoder):
+    def default(self, o):
+        if isinstance(o, Decimal):
+            return float(o)
+
+        return super().default(o)
 
 
 def strip_column_prefix(d):
@@ -10,5 +22,5 @@ def strip_column_prefix(d):
     return res
 
 
-def slugger(name: str):
+def slugger(name: str) -> str:
     return custom_slugify(name)
